fix ut error

+++ pkg/vm/engine/tae/blockio/writer_test.go
@@ -78,8 +78,8 @@ func TestWriter_WriteBlockAndZoneMap(t *testing.T) {
 	require.False(t, res)
 	require.Equal(t, int32(0), zm.GetMin())
 	require.Equal(t, int32(39999), zm.GetMax())
-	sum := int32(0)
-	for i := int32(0); i < 40000; i++ {
+	sum := int64(0)
+	for i := int64(0); i < 40000; i++ {
 		sum += i
 	}
 	require.Equal(t, sum, zm.GetSum())
@@ -114,8 +114,8 @@ func TestWriter_WriteBlockAndZoneMap(t *testing.T) {
 	require.True(t, zm.Contains(int32(40000)))
 	require.True(t, zm.Contains(int32(79999)))
 	require.False(t, zm.Contains(int32(80000)))
-	sum = int32(0)
-	for i := int32(40000); i < 80000; i++ {
+	sum = int64(0)
+	for i := int64(40000); i < 80000; i++ {
 		sum += i
 		if sum < 0 {
 			sum = 0
